Create ZUIPublicFooter and use it.
