+++ src/features/home/layouts/HomeLayout.tsx
@@ -8,11 +8,9 @@ import { useMessages } from 'core/i18n';
 import messageIds from '../l10n/messageIds';
 import ZUIPersonAvatar from 'zui/components/ZUIPersonAvatar';
 import useUser from 'core/hooks/useUser';
-import ZUILogo from 'zui/ZUILogo';
-import { useEnv } from 'core/hooks';
 import ZUITabbedNavBar from 'zui/components/ZUITabbedNavBar';
 import ZUIText from 'zui/components/ZUIText';
-import ZUILink from 'zui/components/ZUILink';
+import ZUIPublicFooter from 'zui/components/ZUIPublicFooter';
 
 type Props = {
   children: ReactNode;
@@ -21,7 +19,6 @@ type Props = {
 
 const HomeLayout: FC<Props> = ({ children, title }) => {
   const messages = useMessages(messageIds);
-  const env = useEnv();
 
   const path = usePathname();
   const lastSegment = path?.split('/').pop() ?? 'home';
@@ -84,26 +81,7 @@ const HomeLayout: FC<Props> = ({ children, title }) => {
         />
       </Box>
       <Box minHeight="90dvh">{children}</Box>
-      <Box
-        alignItems="center"
-        component="footer"
-        display="flex"
-        flexDirection="column"
-        mx={1}
-        my={2}
-        sx={{ opacity: 0.75 }}
-      >
-        <ZUILogo />
-        <ZUIText variant="bodySmRegular">Zetkin</ZUIText>
-        <ZUILink
-          href={
-            env.vars.ZETKIN_PRIVACY_POLICY_LINK ||
-            'https://www.zetkin.org/privacy'
-          }
-          size="small"
-          text={messages.footer.privacyPolicy()}
-        />
-      </Box>
+      <ZUIPublicFooter />
     </Box>
   );
 };
